ChannelServer>Channel, ServerPath added, ChannelPath added
ServerPath is a revision of the renamed ClientPath. Its purpose is to be
abstract. Channel became static and all functionality went into
ChannelPath for the better. I will be doing this with the other servers.
It was a frequent layout I didn't want to mess with. It is more
organized this way too. More static, less needless instantiation.