renamed LoginServerGUI to ServerGUI, disposed of UpdateClientCountTask
I wanted UpdateClientCountTask extends TimerTask to be a functional
interface, alas it isn't. It is better looking and better placed in the
GUI, as it has nothing to do with the other project. If I really wanted
to I would just create another class inside this one, which is seemingly
useless. If I ever need more tasks or the like it may just be a good
idea to create a new function that sets it up.