started work on a new gui, which features much, much more
I want configuration files to be present. I want a tabbed experience in
that each channel will be displayed and the amount of channels open
reflect that of the MySQL database. There was no scrolling in previous
version and I am not familiar with the window build system the original
creator used to manually change it. I am, however, able to use other
ones and I can use the standard border layout JFrame for this one. In
fact, everything right now is border layout. Probably will be until I
get deeper with the config windows, etc. That should be absolute.
Another important thing I should of noted earlier in the paragraph, is I
wanted to make it so everything leeches a database. This makes it so
your buddy can host a login server, too. Which, in the real world, would
allow multiple login points.